Conflicts:
	Makefile

+++ scripts/ci/common.sh
@@ -17,7 +17,10 @@ function runTests() {
 }
 
 if [ -n "${APM_SERVER_BRANCH}" ]; then
- BUILD_OPTS="${BUILD_OPTS} --apm-server-build https://github.com/elastic/apm-server.git@${APM_SERVER_BRANCH}"
+ APM_SERVER_BRANCH_VERSION=${APM_SERVER_BRANCH%;*}
+ APM_SERVER_BRANCH_TYPE=${APM_SERVER_BRANCH//$APM_SERVER_BRANCH_VERSION/}
+ APM_SERVER_BRANCH_TYPE=${APM_SERVER_BRANCH_TYPE//;/}
+ BUILD_OPTS="${BUILD_OPTS} --apm-server-build https://github.com/elastic/apm-server.git@${APM_SERVER_BRANCH_VERSION} ${APM_SERVER_BRANCH_TYPE} "
 fi
 
 if [ -z "${DISABLE_BUILD_PARALLEL}" ]; then
@@ -29,5 +32,5 @@ ELASTIC_STACK_VERSION=${ELASTIC_STACK_VERSION:-'master'}
 # assume we're under CI if BUILD_NUMBER is set
 if [ -n "${BUILD_NUMBER}" ]; then
   # kill any running containers under CI
-  docker ps -aq | xargs -t docker rm -f || true
+  [ -n "$(docker ps -aq)" ] && docker ps -aq | xargs -t docker rm -f || true
 fi
